refactor: extract definition ranking helpers
