+++ src/indexer/call-graph-constants.ts
@@ -1,0 +1,47 @@
+export const CALL_GRAPH_SYMBOL_CHUNK_TYPES = new Set([
+  "function_declaration",
+  "function",
+  "arrow_function",
+  "export_statement",
+  "method_definition",
+  "class_declaration",
+  "interface_declaration",
+  "type_alias_declaration",
+  "enum_declaration",
+  "function_definition",
+  "class_definition",
+  "class_specifier",
+  "struct_specifier",
+  "namespace_definition",
+  "decorated_definition",
+  "method_declaration",
+  "type_declaration",
+  "type_spec",
+  "function_item",
+  "impl_item",
+  "struct_item",
+  "enum_item",
+  "trait_item",
+  "mod_item",
+  "trait_declaration",
+  "trigger_declaration",
+  "test_declaration",
+  "struct_declaration",
+  "union_declaration",
+  // Synthetic Swift declarations or declarations specific to tree-sitter-swift.
+  "actor_declaration",
+  "extension_declaration",
+  "protocol_declaration",
+  "protocol_function_declaration",
+  "init_declaration",
+  "deinit_declaration",
+  "subscript_declaration",
+  // GDScript declarations whose names participate in the call graph.
+  // `function_definition` and `class_definition` are already in the set
+  // above (shared with Python/C/Bash and Python, respectively).
+  "constructor_definition",
+  "enum_definition",
+  "signal_statement",
+  "const_statement",
+  "class_name_statement",
+]);
